Dummy commit: s/localhost/127.0.0.1/
As we have added a new stream to the configuration, we have hit a bug in
Scap3 that now prevents us from deploying the service. This is a dummy
commit to unblock us. It is safe to keep the commit in or revert it at
any point, since it doesn't change anything in the deploy repo.

Bug: T167670
Bug: T169141
Change-Id: I7941c40bb6dd0b79d6e04ec0498283d9decabfe1