rdkafka: Set the maximum buffer size to 65Mb
By default, librdkafka fetches all of the messages in a topic up to
queued.max.messages.kbytes, which by default it set to 1Gb. If a client
is behind the tip of the queue by a lot, EventStreams will quickly grow
in memory, so set the buffer size to a sensible amount.

Bug: T199813
Change-Id: I263dc5f71219b4e760fa5bd2fd7d4c6b7e5331ef